1. Changed TransactionItem to use Big Decimal. 2. Changed Double to Big Decimal in PreferenceOption. 3. Updated remaining double variables to BigDecimal

+++ auth/presentation/src/main/java/com/ravikantsharma/auth/presentation/user_preference/OnboardingPreferencesViewState.kt
@@ -5,6 +5,7 @@ import com.ravikantsharma.core.domain.model.Currency
 import com.ravikantsharma.core.domain.model.DecimalSeparator
 import com.ravikantsharma.core.domain.model.ExpenseFormat
 import com.ravikantsharma.core.domain.model.ThousandsSeparator
+import java.math.BigDecimal
 
 data class OnboardingPreferencesViewState(
     val expenseFormat: ExpenseFormat = ExpenseFormat.MINUS_PREFIX,
@@ -12,7 +13,7 @@ data class OnboardingPreferencesViewState(
     val decimalSeparator: DecimalSeparator = DecimalSeparator.DOT,
     val thousandsSeparator: ThousandsSeparator = ThousandsSeparator.COMMA,
     val enableStartTracking: Boolean = true,
-    val amount: Double = -10382.45,
+    val amount: BigDecimal = BigDecimal(-10382.45),
     val exampleFormat: String = NumberFormatter.formatAmount(
         amount = amount,
         expenseFormat = expenseFormat,
